fix(letters): repair structured editing build and nested list capture
cargo clippy -D warnings fails the CI test job, so the widened ODT
automatic paragraph tuple becomes a named struct instead of tripping
type_complexity. Also fixes the u32 list level passed to rdocx, the
double mutable borrow of doc.page in the ODT styles pass, and makes
typing over a styled selection keep that style so the structured editor
test holds. Nested list markers are now matched after the render indent
so nesting round-trips through the buffer.

+++ letters-core/src/docx.rs
@@ -177,8 +177,8 @@ pub fn write(doc: &Document, path: &str) -> Result<(), String> {
         let para = &paras[i];
         i += 1;
         let mut p = match para.style.list {
-            ListKind::Bullet => out.add_bullet_list_item("", para.style.list_level as usize),
-            ListKind::Numbered => out.add_numbered_list_item("", para.style.list_level as usize),
+            ListKind::Bullet => out.add_bullet_list_item("", u32::from(para.style.list_level)),
+            ListKind::Numbered => out.add_numbered_list_item("", u32::from(para.style.list_level)),
             ListKind::None => out.add_paragraph(""),
         };
         if let Some(level) = para.style.heading {
